requirements.txt
requests==2.32.3
pyspark==3.5.5
delta-spark==3.2.1
apache-airflow==2.10.5
apache-airflow-providers-postgres==5.13.1
